Add Support for Square Tiles
This commit updates the client side javascript to support square grid
tiles in addition to hex tiles. This change involved leaning on the
orientation class a little more heavily so that it could calculate
position the same across shapes. This makes the code much easier to work
with but also more clever. We should eventually refactor the orientation
class to be provided rather than stored.