Enable Map Auto-Save on events
This commit adds a temporary method to perform ajax requests to the
server to update the map whenever there is a change. In order to make
map updates simpler to handle this commit changes tiles on maps to be a
text field that stores JSON since we will always want to save the entire
map when a change is made since the client is king and the server
doesn't need to track transactions.